feat: restructure dwx adapters
